requirements.txt
python-amazon-simple-product-api
click
google-api-python-client
mock
pytest
boto3
